Moved the command tests
